fsd starts with real NSConnector that sends requests to NS.
