+++ testing.go
@@ -51,18 +51,6 @@ func (s *InMemoryChunkStorage) Exists(id string) (exists bool) {
     return
 }
 
-type SpyNSConnector struct {
-    receivedChunks []string
-}
-
-func (c *SpyNSConnector) ReceivedChunk(id string) {
-    c.receivedChunks = append(c.receivedChunks, id)
-}
-
-func (c *SpyNSConnector) Reset() {
-    c.receivedChunks = nil
-}
-
 
 func NewGetChunkRequest(id, token string) *http.Request {
     req, _ := http.NewRequest(http.MethodGet, fmt.Sprintf("/chunks/%s?token=%s", id, token), nil)
